+++ tests/Serializers/RBXTest.php
@@ -5,15 +5,19 @@ namespace Rubix\ML\Tests\Persisters\Serializers;
 use Rubix\ML\Encoding;
 use Rubix\ML\Persistable;
 use Rubix\ML\Classifiers\DummyClassifier;
-use Rubix\ML\Persisters\Serializers\Gzip;
-use Rubix\ML\Persisters\Serializers\Serializer;
+use Rubix\ML\Serializers\RBX;
+use Rubix\ML\Serializers\Serializer;
 use PHPUnit\Framework\TestCase;
+use Rubix\ML\Exceptions\RuntimeException;
+use stdClass;
+
+use function serialize;
 
 /**
  * @group Serializers
- * @covers \Rubix\ML\Persisters\Serializers\Gzip
+ * @covers \Rubix\ML\Serializers\RBX
  */
-class GzipTest extends TestCase
+class RBXTest extends TestCase
 {
     /**
      * @var \Rubix\ML\Persistable
@@ -21,7 +25,7 @@ class GzipTest extends TestCase
     protected $persistable;
 
     /**
-     * @var \Rubix\ML\Persisters\Serializers\Gzip
+     * @var \Rubix\ML\Serializers\RBX
      */
     protected $serializer;
 
@@ -30,7 +34,7 @@ class GzipTest extends TestCase
      */
     protected function setUp() : void
     {
-        $this->serializer = new Gzip(1);
+        $this->serializer = new RBX();
 
         $this->persistable = new DummyClassifier();
     }
@@ -40,7 +44,7 @@ class GzipTest extends TestCase
      */
     public function build() : void
     {
-        $this->assertInstanceOf(Gzip::class, $this->serializer);
+        $this->assertInstanceOf(RBX::class, $this->serializer);
         $this->assertInstanceOf(Serializer::class, $this->serializer);
     }
 
@@ -57,5 +61,32 @@ class GzipTest extends TestCase
 
         $this->assertInstanceOf(DummyClassifier::class, $persistable);
         $this->assertInstanceOf(Persistable::class, $persistable);
+    }
+
+    /**
+     * @return array<mixed>
+     */
+    public function deserializeInvalidData() : array
+    {
+        return [
+            [3],
+            [new stdClass()],
+        ];
+    }
+
+    /**
+     * @test
+     *
+     * @param mixed $obj
+     *
+     * @dataProvider deserializeInvalidData
+     */
+    public function deserializeBadData($obj) : void
+    {
+        $data = new Encoding(serialize($obj));
+
+        $this->expectException(RuntimeException::class);
+
+        $this->serializer->unserialize($data);
     }
 }