+++ CHANGELOG.md
@@ -21,6 +21,11 @@
     - Promoted the Other namespace
     - Moved Flysystem persister to the Extras package
     - Change order of Loda hyper-parameters
+    - Persistent Model now accepts an optional serializer
+    - Peristers no longer interact directly with Persistables
+    - Remove Wrapper interface
+    - RBX serializer now accepts base Gzip parameter
+    - Gzip serializer no longer accepts base serializer
 
 - 0.4.1
     - Optimized CART node splitting for low variance continuous features